modelのstructのみ分離
