+++ server/app/models/fgi.go
@@ -1,6 +1,7 @@
 package models
 
 import (
+	"index-indicator-apis/server/app/entity"
 	"index-indicator-apis/server/mysql"
 	"time"
 
@@ -8,36 +9,8 @@ import (
 	"index-indicator-apis/server/fgi"
 )
 
-// Fgi 日足格納
-type Fgi struct {
-	ID        int       `json:"id,omitempty"`
-	CreatedAt time.Time `json:"created_at,omitempty"`
-	NowValue  int       `json:"now_value,omitempty"`
-	NowText   string    `json:"now_text,omitempty"`
-	PcValue   int       `json:"pc_value,omitempty"`
-	PcText    string    `json:"pc_text,omitempty"`
-	OneWValue int       `json:"one_w_value,omitempty"`
-	OneWText  string    `json:"one_w_text,omitempty"`
-	OneMValue int       `json:"one_m_value,omitempty"`
-	OneMText  string    `json:"one_m_text,omitempty"`
-	OneYValue int       `json:"one_y_value,omitempty"`
-	OneYText  string    `json:"one_y_text,omitempty"`
-}
-
-//initFgis マイグレーション
-func initFgis() {
-	var err error
-	db, err := mysql.SQLConnect()
-	if err != nil {
-		panic(err.Error())
-	}
-	defer db.Close()
-	db.AutoMigrate(&Fgi{})
-}
-
 // CreateNewFgis migration後にapiを叩きdbに保存する
 func CreateNewFgis() error {
-	initFgis() //migration
 	fgiClient := fgi.New(config.Config.FgiAPIKey, config.Config.FgiAPIHost)
 	f, err := fgiClient.GetFgi()
 	if err != nil {
@@ -49,7 +22,7 @@ func CreateNewFgis() error {
 		panic(err.Error())
 	}
 	defer db.Close()
-	db.Create(&Fgi{
+	db.Create(&entity.Fgi{
 		CreatedAt: time.Now(),
 		NowValue:  f.Fgi.Current.Value,
 		NowText:   f.Fgi.Current.ValueText,
@@ -67,14 +40,14 @@ func CreateNewFgis() error {
 }
 
 // GetFgis api for webserver
-func GetFgis(limit int) []Fgi {
+func GetFgis(limit int) []entity.Fgi {
 	db, err := mysql.SQLConnect()
 	if err != nil {
 		panic(err.Error())
 	}
 	defer db.Close()
 
-	fgis := []Fgi{}
+	fgis := []entity.Fgi{}
 	db.Order("created_at desc").Limit(limit).Find(&fgis)
 	return fgis
 }
